Address feedback from review
- Removed `collector-spring-native-config.yaml` file, which was a
leftover from the original application, which used the OpenTelemetry
collector.
- Customized the metrics exporter to set the desired aggregation
temporality for the AppSignal collector, fixing histograms.
- Set the AppSignal collector endpoint explicitly in the config,
rather than passing it via environment variables.
- Added an additional span exporter that logs to the console to
visualise the traces in their entirety and debug why the error is
not reported.
- Renamed the app to `spring-native-collector`.
- Added the tests container to `docker-compose.yml` and set it to be
ran on CI.
