--- java/spring-boot/app/spring-native/collector-spring-native-config.yaml
@@ -1,19 +1,0 @@
-receivers:
-  otlp:
-    protocols:
-      http:
-        endpoint: "0.0.0.0:4318"
-exporters:
-  debug:
-    verbosity: detailed
-service:
-  pipelines:
-    metrics:
-      receivers: [otlp]
-      exporters: [debug]
-    traces:
-      receivers: [otlp]
-      exporters: [debug]
-    logs:
-      receivers: [otlp]
-      exporters: [debug]
