+++ java/spring-native-collector/app/gradle.properties
