+++ java/spring-native-collector/app/settings.gradle.kts
