+++ java/spring-native-collector/app/spring-native/src/main/java/io/opentelemetry/example/graal/Controller.java
