+++ java/spring-native-collector/app/spring-native/src/main/resources/application.properties
