+++ java/spring-native-collector/docker-compose.yml
@@ -5,15 +5,14 @@ services:
     volumes:
       - ./app:/app
     environment:
-      OTEL_SERVICE_NAME: "spring-boot"
       OTEL_LOGS_EXPORTER: "none"
-      OTEL_EXPORTER_OTLP_ENDPOINT: "http://appsignal-collector:8099"
-      APPSIGNAL_APP_NAME: java/spring-boot
+      APPSIGNAL_APP_NAME: java/spring-native-collector
       APPSIGNAL_APP_ENV: development
+      SERVER_PORT: "4001"
     env_file:
       - ../../appsignal_key.env
     ports:
-      - "4001:8080"
+      - "4001:4001"
     depends_on:
       - appsignal-collector
   appsignal-collector:
@@ -22,3 +21,13 @@ services:
       - APPSIGNAL_LOG_LEVEL=trace
     env_file:
       - ../../appsignal_key.env
+  tests:
+    image: server-tests
+    build: ../../support/server-tests
+    profiles:
+      - tests
+    depends_on:
+      - app
+    environment:
+      - APP_NAME=java/spring-native-collector
+      - APP_URL=http://app:4001
